+++ client/components/CourseInfo.jsx
@@ -5,7 +5,7 @@ import { useIntl } from "react-intl";
 import { campuses, departments, terms, years } from "../constants";
 import { messages } from "../content/messages/components/CourseInfo";
 
-const CourseInfo = ({
+export const CourseInfo = ({
   errors,
   hasSubmitted,
   setFieldValue,
