+++ client/components/TabSelect.jsx
@@ -1,7 +1,7 @@
 import { Tab, TabList, Tabs, useColorModeValue } from "@chakra-ui/react";
 import React from "react";
 
-const TabSelect = ({ tabs, onChange, selectedTab = 1 }) => {
+export const TabSelect = ({ tabs, onChange, selectedTab = 1 }) => {
   const highlightedColor = useColorModeValue("black", "gray.500");
   const selectedColor = useColorModeValue("green", "blue");
   return (
