+++ client/components/index.js
@@ -1,0 +1,21 @@
+export * from "./AdvancedSearchModal";
+export * from "./Autocomplete";
+export * from "./BanUserModal";
+export * from "./Card";
+export * from "./ChatInfo";
+export * from "./CheckPermissions";
+export * from "./Contributors";
+export * from "./CourseInfo";
+export * from "./Emoji";
+export * from "./Footer";
+export * from "./LinkFields";
+export * from "./LinkIconBar";
+export * from "./LoginForm";
+export * from "./NavBar";
+export * from "./RegisterForm";
+export * from "./RequestsList";
+export * from "./SectionContainer";
+export * from "./SharedChatFields";
+export * from "./TabSelect";
+export * from "./UsersList";
+export * from "./VerifyEmailAlert";
