+++ client/constants/config.js
@@ -1,0 +1,58 @@
+export const campuses = ["UTM", "UTSG", "UTSC"];
+export const statuses = {
+  approved: "approved",
+  pending: "pending",
+  rejected: "rejected",
+};
+export const userStatuses = { admin: "admin", user: "user", banned: "banned" };
+export const departments = [
+  "ANT",
+  "AST",
+  "BIO",
+  "HSC",
+  "CCT",
+  "CSC",
+  "CHM",
+  "CIN",
+  "CLA",
+  "ECO",
+  "ENG",
+  "ENV",
+  "FSC",
+  "GGR",
+  "HIS",
+  "ITA",
+  "MAT",
+  "MGD",
+  "MGM",
+  "MGT",
+  "PHL",
+  "POL",
+  "PSY",
+  "RLG",
+  "SOC",
+  "STA",
+  "WGS",
+  "WRI",
+];
+export const terms = ["Fall", "Winter", "Summer", "Year", "N/A"];
+export const utscLevels = ["A", "B", "C", "D"];
+export const years = ["2021-2022", "2022-2023", "2023-2024", "N/A"];
+export const numToCampus = {
+  1: "UTSG",
+  3: "UTSC",
+  5: "UTM",
+};
+export const letterToTerm = {
+  F: "Fall",
+  S: "Winter",
+  Y: "Year",
+};
+export const supportedLinks = {
+  discord: "http://discord.gg/",
+  whatsapp: "http://chat.whatsapp.com/",
+  slack: "https://join.slack.com/t/",
+  linktree: "http://linktr.ee/",
+};
+export const MIN_NUM_LINKS = 1;
+export const MAX_NUM_LINKS = 4;
