+++ client/gql/index.js
@@ -1,29 +1,4 @@
-import { gql } from "@apollo/client";
-
-export const GET_ADMIN_DATA = gql`
-  query getAdminData(
-    $status1: String!
-    $status2: String!
-    $limit: Float!
-    $userStatus: String!
-  ) {
-    pendingChats: getGroupChatByStatus(status: $status1) {
-      id
-      name
-    }
-    rejectedChats: getGroupChatByStatus(status: $status2) {
-      id
-      name
-    }
-    users: getUsers(limit: $limit) {
-      email
-    }
-    bannedUsers: getUsers(status: $userStatus) {
-      email
-    }
-  }
-`;
-
-export default {
-  GET_ADMIN_DATA,
-};
+export * from "./Admin";
+export * from "./Auth";
+export * from "./GroupChat";
+export * from "./User";
