+++ client/helpers/index.js
@@ -1,16 +1,5 @@
-export function removeDuplicates(arr) {
-  return [...new Set(arr)];
-}
-
-export function mapAsOption(items, field) {
-  return items.map((item) => ({ label: item[field], value: item[field] }));
-}
-
-export const redirect = (path, push, locale, defaultLocale) => {
-  push(`${locale !== defaultLocale ? locale : ""}${path}`);
-};
-
-export default {
-  mapAsOption,
-  removeDuplicates,
-};
+export * from "./formatters";
+export * from "./groupchats";
+export * from "./network";
+export * from "./testing";
+export * from "./utility";
