+++ client/package.json
@@ -19,6 +19,7 @@
     "@emotion/styled": "^11.1.5",
     "axios": "^0.21.1",
     "cookie": "^0.4.1",
+    "cross-fetch": "^3.1.4",
     "dotenv": "^9.0.0",
     "eslint-config-prettier": "^7.2.0",
     "eslint-plugin-prettier": "^3.3.1",
