+++ client/requests/index.js
@@ -1,20 +1,5 @@
-import client from "../apollo-client";
-import { statuses, userStatuses } from "../constants";
-import { GET_ADMIN_DATA } from "../gql";
-
-export async function getAdminData() {
-  const { data: adminData } = await client.query({
-    query: GET_ADMIN_DATA,
-    variables: {
-      status1: statuses.pending,
-      status2: statuses.rejected,
-      limit: 50,
-      userStatus: userStatuses.banned,
-    },
-  });
-  return adminData;
-}
-
-export default {
-  getAdminData,
-};
+export * from "./admin";
+export * from "./auth";
+export * from "./groupChats";
+export * from "./permissions";
+export * from "./users";
