+++ client/requests/users.js
@@ -1,6 +1,6 @@
 import client from "../apollo-client";
 import { userStatuses } from "../constants";
-import { SEARCH_USERS, UPDATE_USER } from "../gql/User";
+import { SEARCH_USERS, UPDATE_USER } from "../gql";
 
 export async function banUser(selectedUser) {
   const { data } = await client.mutate({
